% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maczic.R
\docType{package}
\name{maczic}
\alias{maczic}
\title{maczic: Mediation Analysis for Count and Zero-inflated Count Data.}
\description{
Performs causal mediation analysis for count and zero-inflated count data;
calculates power of causal mediation effects, direct effects, total effects;
implements Instrumental Variable (IV) method to estimate the controlled
(natural) direct and mediation effect, and the bootstrap CIs; calculates
power of controlled (natural) direct and mediation effect using IV method;
performs sensitivity analysis when there is a treatment-induced
mediator-outcome confounder by varing coefficient of treatment in the
confounder model and plots sensitivity graph.
}
